update: improve MyMatrix, begin MultiPCMultiThread

+++ MyMatrix/try/test_row_column.cpp
@@ -10,15 +10,15 @@ int main(int argc, char const *argv[])
     double t;
 
     data = new int *[size];
-    for (size_t i = 0; i < size; i++)
+    for (int i = 0; i < size; i++)
     {
         data[i] = new int[size];
     }
 
     start = clock();
-    for (size_t i = 0; i < size; i++)
+    for (int i = 0; i < size; i++)
     {
-        for (size_t j = 0; j < size; j++)
+        for (int j = 0; j < size; j++)
         {
             data[i][j] = 100;
         }
@@ -28,9 +28,9 @@ int main(int argc, char const *argv[])
     cout << t << endl;
 
     start = clock();
-    for (size_t j = 0; j < size; j++)
+    for (int j = 0; j < size; j++)
     {
-        for (size_t i = 0; i < size; i++)
+        for (int i = 0; i < size; i++)
         {
             data[i][j] = 100;
         }
